Fix some warnings and runtime tests on linux.

+++ Components/MeshLodGenerator/src/OgreLodConfigSerializer.cpp
@@ -250,7 +250,7 @@ namespace Ogre
 	void LodConfigSerializer::writeLodLevels()
 	{
 		writeChunkHeader(LCCID_LOD_LEVELS, calcLodLevelsSize());
-		uint32 size = mLodConfig->levels.size();
+		uint32 size = static_cast<uint32>(mLodConfig->levels.size());
 		writeInts(&size, 1);
 
 		LodConfig::LodLevelList::iterator it = mLodConfig->levels.begin();
@@ -322,7 +322,7 @@ namespace Ogre
 			return;
 		}
 		writeChunkHeader(LCCID_PROFILE, calcLodProfileSize());
-		uint32 size = mLodConfig->advanced.profile.size();
+		uint32 size = static_cast<uint32>(mLodConfig->advanced.profile.size());
 		writeInts(&size, 1);
 		LodProfile::iterator it = mLodConfig->advanced.profile.begin();
 		LodProfile::iterator itEnd = mLodConfig->advanced.profile.end();
