+++ Samples/MeshLod/src/MeshLod.cpp
@@ -455,7 +455,7 @@ void Sample_MeshLod::addToProfile( Real cost )
 		collapser->_getLastVertexCollapseTo(data.get(), pv.dst);
 		// Prevent duplicates if you edit the same vertex twice.
 		size_t size = mLodConfig.advanced.profile.size();
-		for(size_t i=0;i<size;i++){
+		for(uint i=0;i<size;i++){
 			ProfiledEdge& v = mLodConfig.advanced.profile[i];
 			if(v.src == pv.src && v.dst == pv.dst){
 				v.cost = cost;
